Improve the run method
Got rid of the redundant .run_iter() method from GPT. Instead, .run()
checks if the input is a list and calls itself with each input.
Changed extension keyword to format_ keyword to not rely on gpt guessing
the format from the target extension. For .tif, it guessed GeoTIFF+XML,
and dumped product metadata next to the images themselves.
Added an option to add a prefix to the output.