Add additional sources to operators
Added a possibility for operators to define additional sources. Each
operator now has a list of additional sources, where each source is a
dictionary with keys `lxml_element`, 'name', and 'value'. When a node is
added, Graph now checks if the operator has any additional sources. If
it does, the Graph appends them to the <sources> and stores the name and
value of a source in a dictionary. GPT then checks if a graph has any
additional sources and appends them to the call to gpt.

This is probably not the cleanest way to handle this. Maybe operators
should generate the hole <node>, not just their <parameters>? But then
they still need to pass the sources through the graph to gpt, so it
wouldn't change things too much. Time will tell.